requirements.txt
django==3.1.7
djangorestframework==3.12.4
pydash==4.9.3
opencv-python==4.5.1.48
cmake==3.18.4.post1
face-recognition==1.3.0
django-cors-headers==3.7.0
psycopg2-binary==2.8.5